Fix Android build.

+++ InternetMap/MapController.cpp
@@ -190,8 +190,8 @@ void MapController::highlightConnections(NodePointer node) {
         }
     }
     
-    static const unsigned long NUM_IMPORTANT_CONNECTIONS = 50;
-    static const unsigned long NUM_RENDERED_CONNECTIONS = 100;
+    static const unsigned int NUM_IMPORTANT_CONNECTIONS = 50;
+    static const unsigned int NUM_RENDERED_CONNECTIONS = 100;
     
     if (filteredConnections.size() > NUM_RENDERED_CONNECTIONS) {
         // show a few of the most important ones, then a random selection from the rest
@@ -205,7 +205,7 @@ void MapController::highlightConnections(NodePointer node) {
     Color brightColor = ColorFromRGB(SELECTED_CONNECTION_COLOR_BRIGHT_HEX);
     Color dimColor = ColorFromRGB(SELECTED_CONNECTION_COLOR_DIM_HEX);
 
-    for(unsigned int i = 0; i < std::min(filteredConnections.size(), NUM_RENDERED_CONNECTIONS); i++) {
+    for(unsigned int i = 0; i < std::min((unsigned int)(filteredConnections.size()), NUM_RENDERED_CONNECTIONS); i++) {
         ConnectionPointer connection = filteredConnections[i];
         NodePointer a = connection->first;
         NodePointer b = connection->second;
